fix: исправлен reviewService в ReviewController, userful - Integer, getEvenFeed для безопасного исполнения запроса и избежания SQL-инъекций jdbcTemplate.query() заменён на jdbcTemplate.queryForList(). В checkUserExists теперь запрос сделан через jdbcTemplate.queryForObject(), addEventToFeed теперь проверяет наличие userId в таблице

+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -20,5 +20,5 @@ public class Review {
     private Long userId;
     @NotNull(message = "id фильма должен быть задан")
     private Long filmId;
-    private Long useful; //кол-во лайков обзора
+    private Integer useful; //кол-во лайков обзора
 }
